better handling of invalid UTF-8
Also, be smarter about when to generate the
"Task failed, not executing subsequent tasks" message from
the all task (in particular, it's not necessary if the
very last task fails).

Also, be smarter about generating log output in the copy
task: log messages should be printed to stdout even if we
don't want them to be visible to the student.